requirements.txt
amqp==2.5.1
arrow==0.15.4
anyjson==0.3.3
bleach==1.4.2
certifi==2018.1.18
chardet==3.0.4
decorator==4.4.0
defusedxml==0.5.0
dicttoxml==1.6.6
dj-database-url==0.3.0
dj-static==0.0.6
Django==2.2
django-autocomplete-light==2.3.1
django-cors-headers==3.0.2
django-crispy-forms==1.4.0
django-extensions==1.5.9
django-filter==2.2.0
django-guardian==1.3
django-toolbelt==0.0.1
djangorestframework==3.10.0
djangorestframework-simplejwt==4.3.0
drf-nested-routers==0.91
future==0.17.1
gunicorn==19.3.0
html5lib==0.9999999
idna==2.8
importlib-metadata==0.19
isoweek==1.3.0
jsonpickle==0.9.3
lxml==4.3.3
mock==3.0.5
networkx==2.2
oauthlib==3.0.1
psycopg2==2.7.1
PyJWT==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2016.10
requests==2.22.0
requests-oauthlib==1.2.0
six==1.10.0
sqlparse==0.3.0
static3==0.7.0
Unidecode==1.0.23
urllib3==1.24.1
vine==1.3.0
xmltodict==0.12.0
zipp==0.5.2
